feat: rename app
