+++ android/app/build.gradle
@@ -85,9 +85,9 @@ android {
     buildToolsVersion rootProject.ext.buildToolsVersion
     compileSdk rootProject.ext.compileSdkVersion
 
-    namespace "com.reactnativetemplate"
+    namespace "com.app.synced"
     defaultConfig {
-        applicationId "com.reactnativetemplate"
+        applicationId "com.app.synced"
         minSdkVersion rootProject.ext.minSdkVersion
         targetSdkVersion rootProject.ext.targetSdkVersion
         versionCode 1
