+++ android/app/src/main/java/com/app/synced/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.reactnativetemplate
+package com.app.synced
 
 import com.facebook.react.ReactActivity
 import com.facebook.react.ReactActivityDelegate
@@ -14,7 +14,7 @@ class MainActivity : ReactActivity() {
    * Returns the name of the main component registered from JavaScript. This is used to schedule
    * rendering of the component.
    */
-  override fun getMainComponentName(): String = "ReactNativeTemplate"
+  override fun getMainComponentName(): String = "Synced"
 
   /**
    * Returns the instance of the [ReactActivityDelegate]. We use [DefaultReactActivityDelegate]
