+++ ios/Synced.xcodeproj/xcshareddata/xcschemes/Synced.xcscheme
@@ -15,9 +15,9 @@
                   <BuildableReference
                      BuildableIdentifier = "primary"
                      BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-                     BuildableName = "ReactNativeTemplate.app"
-                     BlueprintName = "ReactNativeTemplate"
-                     ReferencedContainer = "container:ReactNativeTemplate.xcodeproj">
+                     BuildableName = "Synced.app"
+                     BlueprintName = "Synced"
+                     ReferencedContainer = "container:Synced.xcodeproj">
                   </BuildableReference>
                </EnvironmentBuildable>
             </ActionContent>
@@ -33,9 +33,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-               BuildableName = "ReactNativeTemplate.app"
-               BlueprintName = "ReactNativeTemplate"
-               ReferencedContainer = "container:ReactNativeTemplate.xcodeproj">
+               BuildableName = "Synced.app"
+               BlueprintName = "Synced"
+               ReferencedContainer = "container:Synced.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -51,9 +51,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "00E356ED1AD99517003FC87E"
-               BuildableName = "ReactNativeTemplateTests.xctest"
-               BlueprintName = "ReactNativeTemplateTests"
-               ReferencedContainer = "container:ReactNativeTemplate.xcodeproj">
+               BuildableName = "SyncedTests.xctest"
+               BlueprintName = "SyncedTests"
+               ReferencedContainer = "container:Synced.xcodeproj">
             </BuildableReference>
          </TestableReference>
       </Testables>
@@ -73,9 +73,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "ReactNativeTemplate.app"
-            BlueprintName = "ReactNativeTemplate"
-            ReferencedContainer = "container:ReactNativeTemplate.xcodeproj">
+            BuildableName = "Synced.app"
+            BlueprintName = "Synced"
+            ReferencedContainer = "container:Synced.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </LaunchAction>
@@ -90,9 +90,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "ReactNativeTemplate.app"
-            BlueprintName = "ReactNativeTemplate"
-            ReferencedContainer = "container:ReactNativeTemplate.xcodeproj">
+            BuildableName = "Synced.app"
+            BlueprintName = "Synced"
+            ReferencedContainer = "container:Synced.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </ProfileAction>
