+++ ios/Synced/Colors.xcassets/BootSplashBackground-e626f8.colorset/Contents.json
