+++ ios/Synced/Images.xcassets/AppIcon.appiconset/Contents.json
