+++ ios/Synced/Images.xcassets/BootSplashLogo-e626f8.imageset/logo-e626f8@2x.png
